Stop any PDML generation when a new filter is applied.
Sometimes the PDML generation can take a long time, even to load up 1000
packets. This slow process will hold up the operations channel because the
pcap loader state will not be as expected (the LoadingPdml bit will be
set). This also seems to block the quit button, so fixing this might fix issue