Make it easier to display the interfaces available for capture.
Like tshark, handle the -D switch

$ termshark -D

If this switch is provided, control is passed through to tshark just as if the
user had piped to e.g. less or set --pass-thru=true.

Like tshark, you can listen on an interface by specifying its 1-based index in
the list termshark displays when invoked with -D e.g.

$ termshark -D
1. eth0
2. ham0
3. tun0
...

$ termshark -i 1 icmp

Of course you could also do

$ termshark -D --pass-thru

but that's not so obvious.