Experimental - remove dependency on cygwin tail on Windows.
This makes use of a fork of github.com/hpcloud/tail to build tail support into
termshark, in order to remove the need to depend on a cygwin installation on
Windows. The golang tail binary is experimental, and probably inefficient -
the changes to tail are to support tracking a binary file (a pcap in this
case) rather than a line-oriented file.

If main.tail-command is not set in termshark.toml, the default on Windows will
now be

termshark.exe --tail <pcap-file>

i.e. termshark is re-invoked with a new hidden argument.