I broke packet structure expansion with the mouse.
The widget that represents each node in the packet structure view needs to be
cached because gowid will compare the "ID" of the widget clicked with the "ID"
of the widget selected when the button is released to activate the click, and
in this case the "ID" is provided by the address of the button widget itself
i.e. &button("[+]"). This doesn't affect expansion by hitting enter because
that doesn't require a separate down and up event like a mouse click. I need
to find a way to test this UI logic to avoid more regressions...