Be more forgiving about PDML well-formedness.
This change is to address issue #3. Termshark runs a tshark process with the
-T pdml flag to generate the structure and hex views. When reading in
interface mode, a pcap file is generated under ~/.cache/... by another
termshark-initiated tshark process. The PDML-generating tshark might catch
that pcap file mid-write, and in particular half-way through a packet. All
versions of tshark will report an error, but older versions - maybe around
1.12 - will also fail to close all XML tags in the generated PDML
stream. Termshark considered this an unexpected error and displayed an error
dialog. With this change, that error is ignored. This is safe because the
PDML-reading mechanism is racey by its nature, and it's not expected by the
design that it always reads a pcap successfully.