For setting the terminal line discipline to generate SIGTSTP.
The idea is for termshark to response to ctrl-z in a similar way to emacs when
started with the -nw flag. The original terminal view is restored, and all
processes in the process group are sent SIGTSTP which will typically result in
them stopping work. A job control shell will allow resumption of the group
with fg - then SIGCONT is sent to the group.

termshark does not handle ctrl-z and sent SIGTSTP/SIGSTOP to its own group
because other processes in the pipeline might be running as root (e.g. sudo
tcpdump | ...) and they will not receive the signal. Better to leave it to the
line discipline.