A start on cleanup up the notion of a source of packets.
This distinguishes between files, fifos, pipes and interfaces. The logic in
termshark.go was becoming cluttered and hard to reason with. This looks a bit
java-ish... The interface is provided at the site of the structs, rather than
by the user of the structs. This is intended to be used by the loader, and
then clients of the loader can pass in conforming structs.