A table widget that supports copy-mode.
This widget wraps gowid's table.Widget and, with user-supplied functions
to return a given row of table data, and the entirety of the table
data (in whatever format the user determines), can be used to support
gowid's copy-mode feature. Currently this table supports copying a whole
row, and the entire table. It's easy to imagine also supporting copying
just a single column, but I don't need that yet and haven't implemented
it. Currently, when the termshark user activates copy-mode in such a
table, the row is highlighted, and then this can be widened to the whole
table by hitting cursor-left. See

https://github.com/gcla/termshark/blob/master/docs/UserGuide.md

for more information, under Copy Mode.