Update to latest gowid to address high CPU/RAM usage.
This attempts to address issue #61. The gowid fix skips processing of contiguous
sequences of mouse movement events, attempting to prevent a backlog of
events from resulting in a laggy UI and high CPU usage over a period of
several seconds. The fix is not perfect - better would be to process all
events unless the event queue is determined to be backing up (by testing
its length), but this seems to help a lot as a short term fix.