In the filter widget, hitting enter on a valid expression now submits it.
This is in support of issue #6. Previously, the user had to hit cursor
right to leave the filter widget focus, select apply, then hit
enter. Now, hitting enter directly inside the filter submits it, if the
expression is valid. To do this, the drop-down menu of filter
completions is by default not in focus; this is to prevent is accepting
enter keypresses. To activate it, the user has to hit cursor down
once. Wireshark acts similarly.