Allow the user to override the shell's TERM variable.
It's common for me, when running under tmux or screen, to have a TERM
variable that doesn't capture the full range of colors that my terminal
emulator can display. You can of course set that before running
termshark, but this provides a simple way to override it every time
termshark runs. Set in the config like this:

[main]
  term = "screen-256color"