Fix the broken menu row calculation interface.
Previously, I calculated above/middle/below based on the underlying
list's determination of which widgets would be on screen. But this is
just wrong - the space above/below the menu slider should be determined
by how many items are above/below the slider. Ideally I would take into
account the number of rows used to render each item in the given space,
but this is too expensive and O(n) in the number of widgets in the
table/list.