If TERM is not recognized, termshark's error was incorrect.
Previously, termshark determined that the startup failure was due to a
TERM issue by checking the specific error type and message returned from
tcell. This was more fragile than I intended - tcell's error changed
when its term entry lookup procedure was extended to run infocmp if the
entry was not found in the compiled-in terminal database.