Add a notion of a "current source" context to the PCAP loader.
This has a different purpose from the main context. Now, the main
context is intended to be invoked when the loader is shutting down
permanently. The "current source" context is cancelled when the loader
is told to switch to another pcap or packet source. Other loaders - such
as the forthcoming stream reassembly loader - can depend on the "current
source" context so that if a stream load operation is underway when
termshark switches pcaps, it will be interrupted and stopped correctly.