Remove a spurious error reported occasionally on console.
If you start termshark like this:

cat no-such-file.pcap | termshark -i -

then sometimes there is a fifo error reported on console:

Fifo error: read |0: file already closed

This happens because termshark starts a process to read from the pipe
and save the pcap to a temporary file so that it can be refiltered,
kept, etc - e.g.

dumpcap -i /dev/fd/3 -w ~/.cache/termshark/tmp.pcap

and a tail process to feed this pcap into termshark

tail -f ~/.cache/termshark/tmp.pcap

which feeds into an OS fifo, and finally a process to generate PSML
which reads from the fifo:

tshark -r - -T psml

If the tail process encounters an error, it sets c.fifoError, which is
what leads to this error on console. But this error can be benign
because termshark closes the fifo itself if the PSML process encounters
an error - so the error can be ErrClosed, if the tail process's Read()
happens after the fifo is closed.