Add a callback to alert the pcap loader client when a new source is ready.
A new source would be e.g. the user selecting another pcap from the
"recent" menu. Note that applying a new display filter is not a new
source and doesn't trigger this callback. This will be used by stream
reassembly to clear out cached structures when a new pcap is selected.