Fix some incorrect uses of TrackedGo().
The purpose of this is to ensure the goroutine has terminated at
shutdown (it should, and if it doesn't, termshark will hang and I'll
noticed). But I need to pass the sync.WaitGroup! Because this argument
is ..., the compiler didn't noticed I omitted it. That meant the result
was the same as using a regular goroutine.