Overhaul the way callbacks are issued from the pcap loader.
The original idea was that each callback would have to close a channel
when ready, those potentially invoking some UI change via the scheduler
before unblocking the caller. So this could hold up the pcap loader
goroutine until a UI update happens. But decided to embed callbacks to
save files to the recents list, and they didn't follow the same
interface - closing a channel - because they relied on the embedder type
to do that. To fix this inconsistency, now there's an array type that
implements IUnPack - if that is provided as a callback, then the list is
"unpacked" into a list of callbacks, and each in turn is provided with a
channel to close. The original channel is not closed until all handlers
have either closed their channel, or were found not to meet the
interface required for this particular operation.