When termshark is started with a pipe, ctrl-c did not close the pipe.
If you start termshark like this:

tshark -i wlan1 -w - | termshark

Then termshark moves stdin to descriptor 3 at startup, and creates a
PipeSource for the pcap loader with value "/dev/fd/3". This causes
termshark to start a dumpcap something like this

dumpcap -P -i /dev/fd/3 -w ~/.cache/termshark/pcaps/_dev_fd_3--2019-10-29--22-39-36.pcap

and then the psml reader reads from
~/.cache/termshark/pcaps/_dev_fd_3--2019-10-29--22-39-36.pcap.

If you hit ctrl-c in termshark, the dumpcap command is terminated. But
that's not enough to let the feeding tshark process know it should stop
- to do that, descriptor 3 should be closed. Termshark did that at
shutdown, but it should do that when the read from the current source is
stopped - *if* the source is a PipeSource.