A function to assist in the building of UI structures.
I might factor this out and make it available in gowid. The idea is to
facilitate the building of UI structures declaratively. When building
gowid UIs, I find I often want a "handle" to a piece of the substructure
- like a *holder.Widget the contents of which I want to change in
response to a UI event - but that holder is embedded in a larger
structure, like a column widget or a pile widget. You might prefer to
write (simplified)

columns.NewFixed(
  widget1,
  widget2,
  holder.New(
    text.New("hello"),
  )
  widget3,...
)

to better exhibit the UI structure in the code. But to date you'd have
had to save the holder to a separate variable in order to be able to
manipulate it:

// globally
h *holder.Widget

h = holder.New(text.New("hello"))

columns.NewFixed(
  widget1,
  widget2,
  h,
  widget3,...
)

func somethingelse() {
  ...
  h.SetSubWidget(text.New("bye", app))
  ...
}

Now you can build the UI structure like this:

// globally
h *holder.Widget

columns.NewFixed(
  widget1,
  widget2,
  assignTo(&h, holder.New(
    text.New("hello"),
  ))
  widget3,...
)

The trick is to use reflection to implement *p = p where p is a pointer
to a widget type, but not limited to a specific widget
like *holder.Widget i.e. you can use this on any pointer type. The
function depends on the user not trying to provide a pointer to a widget
of a different type...