Make button widgets "invert" their style when clicked but before release.
Maybe this should be an option in the button widget more
generally... Button widgets don't issue callbacks until the button is
released, and gowid will check to make sure the button over which the
cursor is released lines up with the button over which the cursor is
placed when the mouse button is clicked. Prior to this change, if the
user clicks and holds the mouse button over a button, the UI doesn't
respons in any way. But if you look at the graph example in gowid, you
can see that the buttons do respond when clicked but prior to release:
https://github.com/gcla/gowid/blob/master/examples/gowid-graph/graph.go#L266

This change makes the same thing happen in termshark - but I ought to
consider moving this into the button itself...