This type supports table.IInvertable, so provide the API.
table.IInvertable means the table supports mapping a row identifier back
to a physical table row. Recap - a gowid table maintains a mapping
between a unique row identifier and the widgets which that row
comprises. When rendering the table looks up a physical row (#3) to
logical row to fetch the widgets. That way, if the table is sorted, the
widget just needs to track the new physical -> logical mapping. An
invertable table model can go in the opposite direction - logical ->
physical. This is needed by the new stream reassembly code (forthcoming)
- when the user clicks on a reassembled piece of the stream in the
termshark UI, termshark behind the scenes will set the selected packet
in the packet list to the packet from which the stream payload
originated. In order to do that, it needs to map from the logical packet
number e.g. packet#9 to the physical, possibly sorted, row in the
rendered table (#55).