A simple widget to issue callbacks when focus is lost or gained.
Termshark uses this widget to make a small UX improvement in the new
stream reassembly UI. Prior to this change, if you click in the list of
stream payloads, termshark will open a small dialog to tell you it has
selected packet X in the packet list (hidden until the stream UI is
closed). The user then has to hit enter or click the close
button. That's fine if the user is already navigating the list of stream
chunks because they probably wouldn't click the mouse except to
explicitly select a specific stream chunks. But if the stream UI focus
is in e.g. the buttons at the bottom, the user might click with the
mouse in the list of stream chunks just to select that section of the UI
so that they can start to move up and down through the list with
e.g. the cursor keys. In that case, it's annoying to have to dismiss the
dialog telling the user they have selected a specific packet in the
packet list. One option would be to make that dialog temporary without
the need to be dismissed - like a message in a status bar. But another
option is to distinguish the mouse clicks - (a) a click inside the list
if the list already has focus, versus (b) a click inside the list when
the list does not have focus. In case (b), termshark will then not
display the dialog. This also gave the the chance to implement this
feature which was requested by a gowid user.

Termshark now sets a boolean in the state of the overall stream UI when
the embedded list widget (for the stream chunks) gets focus and loses
focus - this is done via the callbacks provided by this widget.