A widget to highlight sections of a text widget via a regex.
This widget wraps a text widget (in practice it needs that specific type
for now) and is also provided a regex, a style, and an occurrence
count, N. It will then render that text widget with the Nth match of the
regex highlighted with the supplied style (or all if N == -1).

This is used by the forthcoming stream UI in order to provide a simple
search like Wireshark does. The user can type in a search string in an
edit widget at the bottom of the UI, and then clicking a Next button
will highlight successive matches. Once all the matches in one payload
have been displayed, the UI will select the next payload in the list,
thus letting the user walk through them.