Use a native tail command on Windows.
This removes the dependency on the cygwin tail command. A hidden --tail
flag is now available in termshark - when invoked that way, termshark
will simply tail a file using the github.com/hpcloud/tail package. Note
that I have applied a couple of batches to implement a slightly hacky
byte-based tail, rather than the current implementation which is
line-based. Fixes #18.