Make sure all pdml is loaded up if a live capture is interrupted.
This is a bug fix. You can see this sometimes running termshark like
this:

termshark -i eth0

with auto-scrolling on - then hit ctrl-c. Sometimes the pdml and hex
views aren't populated. If you hit cursor up then down, they will then
be populated after a second or two. What happens is that the ctrl-c
mistakenly stopped all tshark proceses - the psml reader, the iface
reader (correct), but also the pdml and pcap readers (incorrect). Since
these only read select regions of the pcap - i.e. packets on demand -
they should be allowed to complete. That part of the load is hidden from
the user. Because they were stopped, and because they are always slower
than the psml load, termshark is able to show psml for later packets,
but sometimes not the struct and hex views (from the pdml). This is more
apparent when auto-scroll keeps the user focused on the newest/latest
packets. It is fixed by moving up and down because termshark will check
its pdml cache and finding the required packets missing, start another
pdml load. The correct fix is to complete the original pdml load.