Make better use of space when navigating stream reassembly ascii.
This removes unnecessary space used above and below the widget for
framing characters - so the widget is framed at the sides only. This
makes it easier to see the client and server conversation when in ascii
mode and allows more to be presented in a given number of terminal
lines.