Extend the PSML table row model to colorize packets if possible.
If color information has been extracted from the PSML, it is used to
colorize the packet list. Note that the color filters used come from
Wireshark - typically /usr/share/wireshark/colorfilters and
~/.wireshark/colorfilters.

If termshark is running in a low-color mode (e.g. TERM=screen with only
8 colors), the rich RGB colors from wireshark are mapped to the
"closest" available color using go-colorful.

If color information is missing for a packet, or not supported by the
user's tshark, packets are colored white or black according to whether
or not dark mode is enabled.