Allow the scrollbar to hide if the data fits in the space available.
This is a new option provided to the scrollbar widget. If enabled, the
scrollbar will simply render its data widget if the data available would
fit in the space provided to the scrollbar. Note that this adjusts the
focus path of certain elements in the UI... This is fragile. I really
need a better way e.g. labelling widgets within containers like columns
or piles, rather than using a numeric index.