Support moving through data via a right-mouse-click on the scrollbar
Widgets that are wrapped with withscrollbar.New(...) are supplied with a
vertical scrollbar on the right. Gowid has been updated so that the
scrollbar now supports right clicks, the idea being that a right click
3/4 of the way down the scrollbar moves directly to 3/4 of the way
through the scrollable data. withscrollbar now processes right clicks on
the menu and passes the positional request through to the scrollable
data directly via a SetPos() call - if the underlying scrollable widget
implements that API.