Preserve helpful message in case of capture failure
If termshark can't capture on a live interface, it prints a message
that's intended to be useful to the user - on Linux, ran setcap against
the capture binary. If the capture binary is termshark in its special
TERMSHARK_CAPTURE_MODE, then just print dumpcap - because only dumpcap
requires setcap, not tshark.