A function to read from an interface with dumpcap then tshark
If dumpcap fails, fall back to tshark. This is intended to provide the
best of both worlds when reading a live capture on Unix - the more
efficient dumpcap if possible, tshark if not.

Note that one consequence of this is that on Unix, there are likely now
two processes started for a live capture - termshark with
TERMSHARK_CAPTURE_MODE=1, which then starts dumpcap as a child. This
means capture termination needs to change to kill both processes.