Show the underlying frame number when a stream segment is clicked
When you reassemble a stream, you can scroll through the individual
segments in the main UI pane - viewing the data as hex, ascii or raw. If
you click on one of these segments, the current UI displays a message
like this:

Selected packet 1234.

But this message is misleading - it's stream segment 1234, not packet
1234. This change fixes that so the message displays the actual frame
number selected, which is more useful. If you hit q, the main termshark
UI will have selected that frame in the packet list view.