Make sure packet capture works under sudo for hybrid capture approach
This branch captures packets by running dumpcap first with the provided
interface, then if that fails, running tshark. This is to exploit the
more efficient - and safe - dumpcap, if possible, and fall back to
tshark e.g. for extcap interfaces. This relies on running termshark
itself as the capture binary under a special environment - essentially

TERMSHARK_CAPTURE_MODE=1 termshark -i randpkt

which termshark then reroutes immediately to dumpcap+tshark.

If run under sudo, it's not enough to use "termshark" as the capture
command, because it might not be found in the PATH used when termshark
starts termshark (itself). So instead, use os.Argv[0] which will be the
absolute path to termshark.