Termshark doesn't shut down correctly under some conditions
Without this fix, here's how to reproduce:

- run termshark with a live interface e.g. eth0
- apply a display filter in the UI e.g. "tcp", hit enter
- hit q, enter - termshark hangs

The problem is down to the way contexts are used. Every load from tshark
builds on the main context - that stays the same throughout the life of
termshark. Each file (e.g. if user picks from recents) uses a "source"
context which can be cancelled by the main one. Reading from an
interface uses the "source" context to control interruption of the
tshark processes. If you apply a filter, the tshark processes that read
the pdml have to be restarted. They reset the "source" context. But, the
interface reader generating the source of the pdml are still using the
old source context. Hitting q cancels the "source" context - but that
leaves the interface reader untouched. Termshark, by design, insists that
everything shuts down cleanly, so blocks, waiting for the interface
reader to stop.

The fix is to cancel the main context on shutdown.