Hacky attempt to abstract requirements for a copy-mode table
A copy-mode table is intended to wrap a table widget and makes its data
available to be copied by the user (e.g. 'c' in the UI, then ctrl-c). I
want to use it in another part of the UI, but I need to provide a
subwidget that is not a table.BoundedWidget. So this is an attempt to
generalize what I need from the sub-widget. It's untidy. These are the
APIs I've relied on so far from a copy-mode table.