Control termshark's cache of pcap files on disk
When termshark is run in live capture mode, it saves the pcaps it
creates under $XDG_CACHE_HOME/termshark/pcaps/. Termshark doesn't yet
cap the growth of that directory. Ross reported his cache dir was 40GB!
This change adds a new config option to put a limit on this directory e.g.

[main]
  disk-cache-size-mb = 300

Termshark will check the disk cache once per session, after the first 5s
of inactivity.

Omit this config setting, or set the value to -1, to preserve the
current behavior and let the directory grow. The default is -1 because I
don't want to suddenly delete pcaps that existing users might expect to
remain there based on termshark's current behavior.