Prevent a race allowing the submission of an invalid filter
Prior to this change, it's possible to type an invalid filter expression
and submit it - using enter - in the brief interval before termshark
determines it's invalid and sets the filter field red.

The fix is to delay the submission of a filter on enter until the filter
expression is determined valid. If the filter changes in the meantime,
then the enter is forgotten. If enter is still pending when the
expression becomes green, then it's submitted.