A widget to provide a vim "last-line" or emacs "minibuffer"
This is intended to be displayed towards the bottom of the screen and
provide a way to issue termshark operations by typing rather than
navigating through the menus. When open, the available commands can be
shown by hitting "tab"; the user can then select from a list. Some
commands may take arguments.

This is step 1 in providing vim-like support, or easier navigation of
termshark using the keyboard.