Not sure how to make viper remove items from the toml!
This is a workaround for now. I had thought this API correctly removed
items from the toml file, but recent testing shows that it doesn't. The
hack is to write an empty string for the value instead; this works as
long as the client of the config doesn't need to distinguish between no
string and the empty string.