Bug fix - these dialogs didn't open!
This is down to a problem in my API for opening dialogs. The minibuffer
is a dialog, and when it closes, it reinstates the view that was present
when it was opened. If another dialog is opened in the meantime, it is
closed as a side effect. So the user never sees these dialogs. A
workaround is to schedule them to open the next time round termshark's
main select loop.