Add some vim-key awareness to termshark!
This is a start, and by no means complete. I have not yet attempted to
make these keys configurable, but could in the future with a "map"
minibuffer command. Here are the keys currently available:

:       - open "lastline" mode
:r      - load a new pcap
:q      - quit if confirmed
:q!     - quit immediately
ZZ      - quit immediately
C-w C-w - cycle through the termshark panes (like tab)
C-w =   - reset the spacing so views are of equal size

In tabular views:

gg      - go to top
G       - go to bottom

These take a numberic prefix to go to a specific line e.g. 27gg