A limited function to build a widget containing scrollable text
This widget constructs a text widget and allows it to be scrolled. But
this widget is limited - it assumes no line will wrap. To make this
happen it ensures that any lines that are too long are clipped. It makes
this assumption because my scrollbar APIs are not well designed, and
functions like ScrollPosition and ScrollLength don't understand the
current rendering context. That means if the app is resized, and a line
now takes two screen lines to render and not one, the scrollbar can't be
built accurately. Until I design a better scrollbar API, this will work
- I'm only using it for limited information dialogs at the moment.