Hide QuitRequestChannel sends behind an API function
This function is allowed to fail to send a quit requested message. But
it's ok if it happens - it means one is already enqueued, and the main
select loop hasn't read it out yet. This would happen in practice if the
flow of control results in two such calls before control returns to the
main loop. This doesn't happen during normal operation, but can with a
new feature I'm adding that allows for vim-style "macros" that result in
a sequence of keypresses being sent to the application in between
renders - for example, if the user's macro results in two key sequences
that would quit termshark.