A widget to translate specific keys into key sequences
This package provides a widget that can map one keypress to a sequence
of keypresses. If the user pnovides as input a key that is mapped, the
sequence of resulting keypresses is played to the subwidget before
control returns. If the key is not mapped, it is passed through as
normal. I'm going to use this to provide a vim-like macro feature in
termshark.