Bug fix for a possible deadlock
If a pcap load is issued while another is in progress, the first pcap
must be stopped. But this callback doesn't close a channel; instead it
adds the close to a callback issued via app.Run. But this can leave the
main select loop blocked in the flow of control from a message on the
operations channel because control will never return to allow the
closure queued up for app.Run to be executed.