A new callback handler for pcap operations
This one captures the end of a pcap load operation; it then checks to
see if a global variable representing the "jump position" is set; if it
is, it sets focus to the right packet. This is to support global marks;
when the mark is picked, a new pcap may be loaded, and when loaded, the
correct packet has to be selected.