Fixes mark commands to work when pcap is filtered and table sorted
When a mark is stored, termshark should store the packet number i.e. the
integer contained in the PSML <section> element. But this might not
correspond to the row in the UI packet list view table because (a) a
display filter might be applied, and (b) the packet list view table
might be sorted. So now, when a mark is saved, termshark will work out
from the currently selected table row what the actual packet number is,
and store that; and when jumping to a packet, termshark will map it back
to a table row so that can be selected in the UI.