Better UX when using h, j, k, and l to navigate
Termshark's UI is built using a collection of composite widgets. Its
main view shows a pile of three packet-display widgets - the list view,
the structure view and the hex view. Gowid's pile widget will move focus
between these e.g. if the up arrow is hit and not consumed by the widget
in focus, it will bubble up to the pile, and the pile will move focus
vertically up by one. In termshark, this is undesirable - if the user is
scrolling up through the packet structure, it's awkward to have
termshark switch focus to the packet list view when the user reaches the
outermost layer of the packet structure and hits cursor up one more
time; better to simply do nothing and keep focus in the packet structure
view. To do that, I insert this widget that swallows unprocessed
navigational keypresses before they bubble up. This change adds the
vim-navigation keys to the list of keypresses swallowed.