Fixes a bug in the minibuffer/cmdline when nothing has been typed
When the minibuffer is first opened, I don't display the completion
options because it looks better to have a single line for typing at the
bottom, and not an immediately cluttered screen. But if the user hits
tab or enter, then the cmdline will show options. But prior to this
change, hitting enter on a selection that is shown after opening the
minibuffer and immediately hitting tab would not complete that
selection.