Return to correct UI view after maximize/minimize
The default UI view shows all three packet windows in a vertical
pile. If you hit '|', you can cycle between different layouts, and
termshark will remember the chosen layout and preserve it across
restarts. But, if you switch to a layout that is not the default, then
hit '\' to maximize a pane, then hit '\' to un-maximize the pane, the
default layout is chosen. Worse, the behavior is inconsistent because
the layout is not preserved, so on restarting termshark, you'll be taken
back to the layout that was active prior to the max/un-max
operation. This change makes sure that termshark returns to the
preferred+configured layout after a max/un-max operation.