Compile-in some simple themes
This uses rakyll/statik to compile the toml files into simple blobs that
are compiled-in to termshark and made available programmatically via the
Golang http FileSystem
interface (https://golang.org/pkg/net/http/#FileSystem). The idea here
is to provide a small set of default themes, but allow them to be
extended via toml files added to ~/.config/termshark/themes/. Real
filesystem themes will take precedence. The themes are organized in an
attempt to be base16-theme compatible.