Allow use of the end key to resume packet auto-scrolling
When wireshark is run on a live interface, the packet auto-scrolling
setting causes wireshark to scroll down and focus on the latest packet,
as new packets arrive. If the user moves around in the packet list view,
to resume auto-scrolling, then can click a button at the top, near the
menus. Termshark also has an auto-scroll option, but prior to this
change, if the user moved around in the packet list view, auto-scrolling
could not be resumed. Now, the user can hit the end key while in the
packet list view, and termshark will resume auto-scrolling on the
current capture.