Overhaul the way I am handling themes
I had misunderstood base16 with my first implementation. The point of a
base16-theme is so that if your terminal runs in 256-color mode, the
first 16 (+ maybe another 8) colors in the 256-color space are remapped
according to the theme's colors, and then the terminal app simply uses
this range of colors to construct its output. Then, automatically, the
terminal app's colors will follow the colors of the shell from which it
was launched. So instead of building a wide range of themes for
termshark, it's easier to support base16, and then lean on the user
making use of projects like this:
https://github.com/chriskempson/base16-shell

Termshark will now support themes in any of these 3 modes:
- 16 color (e.g. TERM=xterm-16color)
- 256 color (e.g. TERM=xterm-256color)
- truecolor (e.g. COLORTERM=truecolor)

I've added support for loading themes from a small compiled-in
database. Termshark will look first in ~/.config/termshark/themes/ and
then failing that, in it's compiled-in database. The default theme
loaded is /themes/default-[mode].toml. A theme is configured via the
cmdline "theme" command which will set this termshark.toml field e.g.

[main]
  theme = "mytheme"

Termshark still has a color-scheme defined programmatically, but in
practice, it will now always be overridden by the default theme, if
no other theme is set.

The default theme exists in two flavors - 16 and 256. There's no
truecolor theme because termshark's theme loader will fall back to the
256-color theme if the truecolor variant can't be loaded. The default
256 color theme's colors are expressed in RGB format, meaning 24-bit
color. If this theme is loaded when termshark is running in a truecolor
environment, then the colors will be represented precisely via truecolor
ANSI color codes emitted to the terminal. If termshark is genuinely
running in a 256-color environment and not truecolor, then the RBG
colors will be interpolated into the 256-color space, and in practice
the match is close enough. Therefore, skipping the truecolor-specific
theme is a simple short-cut.

I expect the main theme used will be base16. This defines termshark's
colors using the first 16 (24) colors in the 256-color space, meaning
termshark will follow the color scheme of the terminal from which it's
launched. I've added sample dracula and solarized (dark) themes too, for
samples - these use RGB color codes that try to follow the respective
schemes' guidelines. But if your shell is already using dracula or
solarized (dark) via base16-shell, then just set your theme to base16
instead.